Extract SettingsUpdateRequest and simplify updateSettings
